siirrä localstorage itemin poisto samaan komponenttiin jossa se lisätään
